add json wrapper

+++ app/src/main/java/wisc/selfdriving/utility/CarControl.java
@@ -7,11 +7,11 @@ package wisc.selfdriving.utility;
 //make order as object in order to send from UDPService to main
 public class CarControl {
 
-    public int speed_;
-    public int rotation_;
+    public double speed_;
+    public double rotation_;
 
     public CarControl() {
-        this.speed_ = 0;
-        this.rotation_ = 5;
+        this.speed_ = 0.0;
+        this.rotation_ = 5.0;
     }
 }
